+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/StoreResponse.java
@@ -205,7 +205,7 @@ public class StoreResponse {
     }
 
     //NOTE: only used for testing purpose to change the response header value
-    private void setHeaderValue(String headerName, String value) {
+    void setHeaderValue(String headerName, String value) {
         if (this.responseHeaderValues == null || this.responseHeaderNames.length != this.responseHeaderValues.length) {
             return;
         }