+++ sdk/postgresqlflexibleserver/azure-resourcemanager-postgresqlflexibleserver/src/main/java/com/azure/resourcemanager/postgresqlflexibleserver/models/Databases.java
@@ -13,11 +13,12 @@ import com.azure.core.util.Context;
  */
 public interface Databases {
     /**
-     * Deletes a database.
+     * Deletes an existing database.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param serverName The name of the server.
-     * @param databaseName The name of the database.
+     * @param databaseName Name of the database (case-sensitive). Exact database names can be retrieved by getting the
+     * list of all existing databases in a server.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
@@ -25,11 +26,12 @@ public interface Databases {
     void delete(String resourceGroupName, String serverName, String databaseName);
 
     /**
-     * Deletes a database.
+     * Deletes an existing database.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param serverName The name of the server.
-     * @param databaseName The name of the database.
+     * @param databaseName Name of the database (case-sensitive). Exact database names can be retrieved by getting the
+     * list of all existing databases in a server.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
@@ -38,47 +40,49 @@ public interface Databases {
     void delete(String resourceGroupName, String serverName, String databaseName, Context context);
 
     /**
-     * Gets information about a database.
+     * Gets information about an existing database.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param serverName The name of the server.
-     * @param databaseName The name of the database.
+     * @param databaseName Name of the database (case-sensitive). Exact database names can be retrieved by getting the
+     * list of all existing databases in a server.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return information about a database along with {@link Response}.
+     * @return information about an existing database along with {@link Response}.
      */
     Response<Database> getWithResponse(String resourceGroupName, String serverName, String databaseName,
         Context context);
 
     /**
-     * Gets information about a database.
+     * Gets information about an existing database.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param serverName The name of the server.
-     * @param databaseName The name of the database.
+     * @param databaseName Name of the database (case-sensitive). Exact database names can be retrieved by getting the
+     * list of all existing databases in a server.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return information about a database.
+     * @return information about an existing database.
      */
     Database get(String resourceGroupName, String serverName, String databaseName);
 
     /**
-     * List all the databases in a given server.
+     * Lists all databases in a server.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param serverName The name of the server.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return a List of databases as paginated response with {@link PagedIterable}.
+     * @return list of all databases in a server as paginated response with {@link PagedIterable}.
      */
     PagedIterable<Database> listByServer(String resourceGroupName, String serverName);
 
     /**
-     * List all the databases in a given server.
+     * Lists all databases in a server.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param serverName The name of the server.
@@ -86,35 +90,35 @@ public interface Databases {
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return a List of databases as paginated response with {@link PagedIterable}.
+     * @return list of all databases in a server as paginated response with {@link PagedIterable}.
      */
     PagedIterable<Database> listByServer(String resourceGroupName, String serverName, Context context);
 
     /**
-     * Gets information about a database.
+     * Gets information about an existing database.
      * 
      * @param id the resource ID.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return information about a database along with {@link Response}.
+     * @return information about an existing database along with {@link Response}.
      */
     Database getById(String id);
 
     /**
-     * Gets information about a database.
+     * Gets information about an existing database.
      * 
      * @param id the resource ID.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return information about a database along with {@link Response}.
+     * @return information about an existing database along with {@link Response}.
      */
     Response<Database> getByIdWithResponse(String id, Context context);
 
     /**
-     * Deletes a database.
+     * Deletes an existing database.
      * 
      * @param id the resource ID.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
@@ -124,7 +128,7 @@ public interface Databases {
     void deleteById(String id);
 
     /**
-     * Deletes a database.
+     * Deletes an existing database.
      * 
      * @param id the resource ID.
      * @param context The context to associate with this operation.